# docs/api-reference/openapi.yaml
...
openapi: 3.1.0
info:
  title: Lithosphere REST API
  version: 1.0.0
  description: |
    Read-only HTTP endpoints consumed by the Lithosphere Explorer frontend.
    Backed by the indexer's PostgreSQL store and the chain's Cosmos/EVM RPCs.

    Base URL in production: `https://makalu.litho.ai/api`.

    The route SET is mirrored from the running Express routers via
    `scripts/list-routes.ts`. A CI job (`openapi-sync-check`) re-runs that
    introspection and fails if the route SET drifts from `paths:` below.
    Shape drift (response bodies, query params) is NOT gated — those are
    enforced by tests, not by this file.
servers:
  - url: https://makalu.litho.ai/api
    description: Makalu testnet
  - url: http://localhost:4000/api
    description: Local dev (docker-compose)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Thanos SIWE session token from /auth/verify, sent as `Authorization: Bearer <token>`.'
  parameters:
...
    AddressPath:
      name: address
      in: path
      required: true
      description: Hex (0x…) or bech32 (`litho1…`) address. Server lowercases hex inputs.
      schema:
        type: string
...
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
...
paths:
...
  /bridge/status/{txHash}:
    get:
      tags: [Bridge]
      summary: Bridge transfer status for an internal bridge tx hash.
      parameters:
        - name: txHash
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Transfer status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JsonObject' }
        '400': { description: Invalid bridge transaction hash. }
  /bridge/signatures/{txHash}:
    get:
      tags: [Bridge]
      summary: Collected validator signatures for a bridge transfer.
      parameters:
        - name: txHash
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Signature list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JsonObject' }
        '400': { description: Invalid bridge transaction hash. }
  /bridge/transactions/{address}:
    get:
      tags: [Bridge]
      summary: Bridge transfer history for an address.
      parameters:
        - $ref: '#/components/parameters/AddressPath'
      responses:
        '200':
          description: Transfer history.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JsonObject' }
        '400': { description: Invalid wallet address. }

  # ───── Sign in with Thanos (SIWE auth) ────
...